% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNetVISOR.R
\name{generateNewValidationEdges}
\alias{generateNewValidationEdges}
\title{Generate Three Panel (Edge) Graph Display}
\usage{
generateNewValidationEdges(paneled = T)
}
\arguments{
\item{TRUE/FALSE}{boolean toggling paneled display. Defaults to T.}
}
\description{
This function generates a 3-panel visualization of the same graph. The first graph (left) depicts the small-world graph itself. The second graph (middle) depicts the same graph from the left, but with predicted noticeability scores overlayed on top. The third graph (right) displays a heatmap visualization to help illustrate the data from the second (middle) graph. This function focuses on edge data only. The related generateNewValidation function provides similar functionality for nodes in graphs.
}
\examples{
generateNewValidation()
}
\keyword{graph,}
\keyword{heatmap}
\keyword{noticeability,}
\keyword{saliency,}
\keyword{visual,}

